Move kernel_artifacts to iris/tracing/ and fix docstring layout
Relocate kernel_artifacts.py from iris package root to iris/tracing/
subpackage where it belongs alongside other observability code. Update
all 9 import sites and fix the directory layout docstring to include
the codegen hash subdir.
