+++ iris/ccl/all_to_all.py
@@ -9,7 +9,7 @@ Supports both Triton and Gluon implementations based on config.
 import triton
 import triton.language as tl
 import iris
-from iris.kernel_artifacts import iris_launch
+from iris.tracing.kernel_artifacts import iris_launch
 from .config import Config
 from .utils import chiplet_transform_chunked, extract_group_info
