+++ iris/ccl/reduce_scatter.py
@@ -9,7 +9,7 @@ Uses the two-shot approach: reduce assigned tiles and store only to own rank.
 import triton
 import triton.language as tl
 import iris
-from iris.kernel_artifacts import iris_launch
+from iris.tracing.kernel_artifacts import iris_launch
 from .config import Config
 from .utils import chiplet_transform_chunked, ReduceOp, extract_group_info
