+++ iris/ops/all_gather_matmul.py
@@ -15,7 +15,7 @@ import triton
 import triton.language as tl
 import iris
 import iris.x
-from iris.kernel_artifacts import iris_launch
+from iris.tracing.kernel_artifacts import iris_launch
 
 from tritonblas.kernels.stages import GemmContext, ScheduleContext
