+++ iris/ops/matmul_reduce_scatter.py
@@ -19,7 +19,7 @@ from .config import FusedConfig
 from .workspace import FusedWorkspace
 import iris
 import iris.x
-from iris.kernel_artifacts import iris_launch
+from iris.tracing.kernel_artifacts import iris_launch
 
 
 @triton.jit()
